Use File URLs, Add Bed Files, Move Bed logic to statistics, reverts some import.meta logic
