+++ packages/iobio-react-components/src/utils/bedFiles/1k_flank_hg38_shuffled1.bed
@@ -1,0 +1,200 @@
+chr1	777650	779650
+chr1	1691724	1693724
+chr1	3610508	3612508
+chr1	4411094	4413094
+chr1	6198553	6200553
+chr1	12355350	12357350
+chr1	20775604	20777604
+chr1	40691086	40693086
+chr1	44806479	44808479
+chr1	61252484	61254484
+chr1	63322355	63324355
+chr1	71046496	71048496
+chr1	83859536	83861536
+chr1	118646960	118648960
+chr1	120515995	120517995
+chr1	146485740	146487740
+chr1	148316777	148318777
+chr1	148401508	148403508
+chr1	155261433	155263433
+chr1	156428429	156430429
+chr1	217630323	217632323
+chr1	235327569	235329569
+chr1	246565460	246567460
+chr10	1021578	1023578
+chr10	43076110	43078110
+chr10	69482824	69484824
+chr10	78178196	78180196
+chr10	121484532	121486532
+chr10	122210685	122212685
+chr11	232683	234683
+chr11	2884771	2886771
+chr11	58727729	58729729
+chr11	59560276	59562276
+chr11	107442031	107444031
+chr11	119084270	119086270
+chr11	119728525	119730525
+chr11	122206655	122208655
+chr11	125062202	125064202
+chr11	127066001	127068001
+chr11	134223216	134225216
+chr12	12225673	12227673
+chr12	25958380	25960380
+chr12	29782942	29784942
+chr12	54339371	54341371
+chr12	55965042	55967042
+chr12	57533530	57535530
+chr12	58102906	58104906
+chr12	88418279	88420279
+chr12	108879350	108881350
+chr12	114734249	114736249
+chr13	20178462	20180462
+chr13	63737853	63739853
+chr13	77326094	77328094
+chr13	78839049	78841049
+chr13	106941561	106943561
+chr14	23070557	23072557
+chr14	25805712	25807712
+chr14	50594081	50596081
+chr14	52872710	52874710
+chr14	61280338	61282338
+chr14	61761419	61763419
+chr14	69546848	69548848
+chr14	70808863	70810863
+chr14	86061958	86063958
+chr14	93908185	93910185
+chr14	95533783	95535783
+chr14	99360083	99362083
+chr15	28831312	28833312
+chr15	30624733	30626733
+chr15	42403780	42405780
+chr15	52790186	52792186
+chr15	67058255	67060255
+chr15	69461992	69463992
+chr15	70663944	70665944
+chr15	73925470	73927470
+chr15	78943996	78945996
+chr15	79282855	79284855
+chr15	101959197	101961197
+chr16	904224	906224
+chr16	1223775	1225775
+chr16	19765230	19767230
+chr16	28949987	28951987
+chr16	29474063	29476063
+chr16	30747550	30749550
+chr16	56657011	56659011
+chr16	69130290	69132290
+chr17	4949189	4951189
+chr17	7613063	7615063
+chr17	7676578	7678578
+chr17	18855298	18857298
+chr17	30599964	30601964
+chr17	32006871	32008871
+chr17	39032108	39034108
+chr17	40053408	40055408
+chr17	43387312	43389312
+chr17	45247354	45249354
+chr17	45489030	45491030
+chr17	50461535	50463535
+chr17	81228340	81230340
+chr17	82751073	82753073
+chr18	8808817	8810817
+chr18	15329507	15331507
+chr18	24724846	24726846
+chr19	15672087	15674087
+chr19	21593644	21595644
+chr19	36265716	36267716
+chr19	47483294	47485294
+chr2	739889	741889
+chr2	10005030	10007030
+chr2	10448892	10450892
+chr2	28459361	28461361
+chr2	28750745	28752745
+chr2	38514658	38516658
+chr2	52296994	52298994
+chr2	53786009	53788009
+chr2	58205236	58207236
+chr2	65055406	65057406
+chr2	109759389	109761389
+chr2	130427176	130429176
+chr2	145571661	145573661
+chr2	169693490	169695490
+chr2	232417944	232419944
+chr20	8830185	8832185
+chr20	36950579	36952579
+chr20	38420671	38422671
+chr20	47483381	47485381
+chr20	50930000	50932000
+chr20	52454619	52456619
+chr20	58514491	58516491
+chr20	58852265	58854265
+chr20	60473650	60475650
+chr21	6669695	6671695
+chr21	10118487	10120487
+chr22	11706170	11708170
+chr22	29267325	29269325
+chr22	32528082	32530082
+chr22	36775256	36777256
+chr22	44195097	44197097
+chr3	31227714	31229714
+chr3	49097524	49099524
+chr3	52218125	52220125
+chr3	119172599	119174599
+chr3	122955514	122957514
+chr3	128051433	128053433
+chr3	158087983	158089983
+chr3	181722192	181724192
+chr3	186824662	186826662
+chr4	1297069	1299069
+chr4	8999747	9001747
+chr4	17582430	17584430
+chr4	37073212	37075212
+chr4	48830062	48832062
+chr4	138308737	138310737
+chr4	168479514	168481514
+chr4	188158922	188160922
+chr4	190020182	190022182
+chr5	76335828	76337828
+chr5	88672724	88674724
+chr5	91313997	91315997
+chr5	141635825	141637825
+chr5	174335254	174337254
+chr6	2501481	2503481
+chr6	27478022	27480022
+chr6	31776272	31778272
+chr6	111258154	111260154
+chr6	112138199	112140199
+chr6	141803927	141805927
+chr6	146026702	146028702
+chr6	169789052	169791052
+chr6	170023174	170025174
+chr7	841920	843920
+chr7	30568821	30570821
+chr7	36148322	36150322
+chr7	64036935	64038935
+chr7	107743154	107745154
+chr7	111090126	111092126
+chr7	135757800	135759800
+chr7	143406813	143408813
+chr7	148697964	148699964
+chr7	151955401	151957401
+chr7	153432864	153434864
+chr8	2126796	2128796
+chr8	17156144	17158144
+chr8	24513855	24515855
+chr8	30656005	30658005
+chr8	47959940	47961940
+chr8	66924541	66926541
+chr8	98943445	98945445
+chr8	100414939	100416939
+chr8	127793566	127795566
+chr8	136788834	136790834
+chr9	3180396	3182396
+chr9	41072712	41074712
+chr9	66694850	66696850
+chr9	94165325	94167325
+chr9	98795555	98797555
+chr9	127801917	127803917
+chr9	131544528	131546528
+chr9	136700837	136702837
+chr9	137019805	137021805
