+++ packages/iobio-react-components/src/utils/bedFiles/1k_flank_hg38_shuffled2.bed
@@ -1,0 +1,200 @@
+chr1	180130	182130
+chr1	1616407	1618407
+chr1	31880763	31882763
+chr1	40734314	40736314
+chr1	42891430	42893430
+chr1	63314986	63316986
+chr1	64562676	64564676
+chr1	90850644	90852644
+chr1	117699242	117701242
+chr1	149016759	149018759
+chr1	151281311	151283311
+chr1	160219934	160221934
+chr1	160795207	160797207
+chr1	193068015	193070015
+chr1	196607411	196609411
+chr1	203428250	203430250
+chr1	203518481	203520481
+chr1	236065910	236067910
+chr10	3257080	3259080
+chr10	4073580	4075580
+chr10	17326109	17328109
+chr10	22847806	22849806
+chr10	26379867	26381867
+chr10	69573966	69575966
+chr10	71004788	71006788
+chr10	73712509	73714509
+chr10	97400534	97402534
+chr10	101817758	101819758
+chr10	114820539	114822539
+chr10	124417193	124419193
+chr11	1663596	1665596
+chr11	7795942	7797942
+chr11	18401534	18403534
+chr11	18404608	18406608
+chr11	65719468	65721468
+chr11	65918723	65920723
+chr11	68459748	68461748
+chr11	76859866	76861866
+chr11	82642445	82644445
+chr11	84272280	84274280
+chr11	90222049	90224049
+chr11	95088718	95090718
+chr11	118623083	118625083
+chr12	866139	868139
+chr12	2676135	2678135
+chr12	26270827	26272827
+chr12	48170589	48172589
+chr12	66803186	66805186
+chr12	73905939	73907939
+chr12	95072503	95074503
+chr12	116877706	116879706
+chr12	133046620	133048620
+chr13	24737647	24739647
+chr13	77074513	77076513
+chr14	30621303	30623303
+chr14	44506327	44508327
+chr14	56311444	56313444
+chr14	60980051	60982051
+chr14	67818778	67820778
+chr14	87255782	87257782
+chr14	101334493	101336493
+chr14	102076472	102078472
+chr15	25437999	25439999
+chr15	41282989	41284989
+chr15	42494708	42496708
+chr15	48046036	48048036
+chr15	55950400	55952400
+chr15	92644598	92646598
+chr15	96284523	96286523
+chr15	96395186	96397186
+chr15	99248518	99250518
+chr16	1239378	1241378
+chr16	4845210	4847210
+chr16	56400775	56402775
+chr16	82034640	82036640
+chr17	713839	715839
+chr17	14845148	14847148
+chr17	19532461	19534461
+chr17	28719704	28721704
+chr17	30567447	30569447
+chr17	36175685	36177685
+chr17	39618669	39620669
+chr17	40053413	40055413
+chr17	50160568	50162568
+chr17	50537754	50539754
+chr17	82097187	82099187
+chr18	14178130	14180130
+chr18	39799296	39801296
+chr18	53567446	53569446
+chr19	15827931	15829931
+chr19	16019482	16021482
+chr19	17870944	17872944
+chr19	37503465	37505465
+chr19	42764662	42766662
+chr19	50417937	50419937
+chr2	6637748	6639748
+chr2	22748276	22750276
+chr2	39123345	39125345
+chr2	43730540	43732540
+chr2	53886316	53888316
+chr2	62295283	62297283
+chr2	62295286	62297286
+chr2	68528637	68530637
+chr2	95401720	95403720
+chr2	96033915	96035915
+chr2	98519048	98521048
+chr2	105640583	105642583
+chr2	122465224	122467224
+chr2	151052608	151054608
+chr2	176464941	176466941
+chr2	191013161	191015161
+chr2	206764605	206766605
+chr2	206811283	206813283
+chr2	216080801	216082801
+chr2	216411413	216413413
+chr2	240819217	240821217
+chr20	20051557	20053557
+chr20	20075488	20077488
+chr20	28601716	28603716
+chr20	35546486	35548486
+chr20	41714666	41716666
+chr20	45406889	45408889
+chr20	45856405	45858405
+chr20	48384298	48386298
+chr20	63247533	63249533
+chr20	63323312	63325312
+chr21	14752469	14754469
+chr21	27637675	27639675
+chr21	33930604	33932604
+chr21	42021031	42023031
+chr22	12096936	12098936
+chr22	27186370	27188370
+chr22	31836237	31838237
+chr22	33435669	33437669
+chr3	48984963	48986963
+chr3	49000473	49002473
+chr3	50566680	50568680
+chr3	64684107	64686107
+chr3	122365762	122367762
+chr3	127201473	127203473
+chr3	129271295	129273295
+chr3	132721078	132723078
+chr3	158109072	158111072
+chr3	165205947	165207947
+chr3	176326027	176328027
+chr3	186439344	186441344
+chr3	191424527	191426527
+chr3	196926811	196928811
+chr4	3120399	3122399
+chr4	3926651	3928651
+chr4	20003635	20005635
+chr4	27735587	27737587
+chr4	48269020	48271020
+chr4	92262343	92264343
+chr4	185050885	185052885
+chr5	16935259	16937259
+chr5	21458486	21460486
+chr5	40410208	40412208
+chr5	53108822	53110822
+chr5	54882602	54884602
+chr5	55776468	55778468
+chr5	76608124	76610124
+chr5	82858995	82860995
+chr5	95553977	95555977
+chr5	132655835	132657835
+chr5	160257252	160259252
+chr6	3911052	3913052
+chr6	29528419	29530419
+chr6	31821310	31823310
+chr6	40377336	40379336
+chr6	112233758	112235758
+chr6	133837153	133839153
+chr6	149647751	149649751
+chr7	876904	878904
+chr7	1438158	1440158
+chr7	7639026	7641026
+chr7	17307426	17309426
+chr7	26200764	26202764
+chr7	38330110	38332110
+chr7	75357975	75359975
+chr7	75358200	75360200
+chr7	100087981	100089981
+chr7	130912496	130914496
+chr8	19776370	19778370
+chr8	66191092	66193092
+chr8	79766615	79768615
+chr8	90645083	90647083
+chr8	100104624	100106624
+chr8	100336519	100338519
+chr8	106059496	106061496
+chr8	142662522	142664522
+chr9	35785132	35787132
+chr9	35791408	35793408
+chr9	64637994	64639994
+chr9	87824401	87826401
+chr9	89177818	89179818
+chr9	92836668	92838668
+chr9	107114783	107116783
+chr9	125746453	125748453
+chr9	137453656	137455656
