+++ packages/iobio-react-components/src/utils/elasticMetadataClient/indexerTools.mts
@@ -22,8 +22,8 @@
 import readline from 'node:readline/promises';
 import { type StatsOutput } from '../iobioTypes.ts';
 import { BamFileExtensions, getFileMetadata } from '../scoreFileTools.mts';
-import { type ElasticSearchResult, type FileDocument } from '../scoreFileTypes.ts';
-import { generateIobioStats, type CompleteCallback } from '../statisticsClient/statisticsTools.mts';
+import { type ElasticSearchResult } from '../scoreFileTypes.ts';
+import { generateIobioStats, getBedFileUrl, type CompleteCallback } from '../statisticsClient/statisticsTools.mts';
 
 /** Base ElasticSearch arguments */
 export type EsConfig = {
@@ -68,19 +68,6 @@ const iobioProperties = JSON.stringify({
 		},
 	},
 });
-
-/** File Strategy & Bed URL */
-const fileStrategies = ['WGS', 'WXS', 'ChipSeq', 'RNA-Seq'];
-type FileStrategyKey = (typeof fileStrategies)[number];
-type DefaultBedUrls = {
-	[K in FileStrategyKey]: string;
-};
-const bedUrls: DefaultBedUrls = {
-	WGS: 'http://localhost:9080/1k_flank_hg38_shuffled1.bed',
-	WXS: 'http://localhost:9080/hg38_Twist_Bioscience_for_Illumina_Exome_2.5.subset.bed',
-	ChipSeq: 'http://localhost:9080/1k_flank_hg38_shuffled2.bed',
-	'RNA-Seq': 'http://localhost:9080/hg38_Twist_Bioscience_for_Illumina_Exome_2.5.subset.bed',
-};
 
 /**
  * Confirm requested index exists, and add Iobio Field Mappings if needed
@@ -153,17 +140,6 @@ export const searchDocument = async ({ index, documentId, esHost, requestOptions
 		throw new Error(`No document found with id ${documentId}`);
 	}
 	return searchResult;
-};
-
-/** Lookup Default Bed File
- * @param elasticDocument File Centric ElasticSearch document
- * @returns bedFileUrl - string
- */
-export const getBedFileUrl = (elasticDocument: FileDocument) => {
-	const fileStrategy = elasticDocument.analysis?.experiment?.experimentalStrategy;
-	const isValidStrategy = fileStrategy && fileStrategies.includes(fileStrategy);
-	const bedFileUrl = isValidStrategy ? bedUrls[fileStrategy] : '';
-	return bedFileUrl;
 };
 
 /**
