+++ packages/iobio-react-components/src/utils/statisticsClient/statisticsTools.mts
@@ -24,6 +24,7 @@ import fs from 'node:fs';
 import readline from 'node:readline/promises';
 import { calculateMeanCoverage, getBamStatistics } from '../iobioTools.mts';
 import { type StatsOutput } from '../iobioTypes.ts';
+import { type FileDocument } from '../scoreFileTypes.ts';
 
 export type CompleteCallback = (stats: StatsOutput) => Promise<void>;
 
@@ -120,4 +121,37 @@ export const statisticsCLI = async () => {
 	readlineInterface.close();
 
 	generateIobioStats({ fileUrl, indexFileUrl, bedFileUrl, enableFileOutput });
+};
+
+/** File Strategy & Bed URL */
+const fileStrategies = ['WGS', 'WXS', 'ChipSeq', 'RNA-Seq'];
+type FileStrategyKey = (typeof fileStrategies)[number];
+type DefaultBedUrls = {
+	[K in FileStrategyKey]: string;
+};
+
+const bedShuffled1Url = new URL('./../bedFiles/1k_flank_hg38_shuffled1.bed', import.meta.url);
+const bedShuffled2Url = new URL('./../bedFiles/1k_flank_hg38_shuffled2.bed', import.meta.url);
+const bedIlluminaUrl = new URL(
+	'./../bedFiles/hg38_Twist_Bioscience_for_Illumina_Exome_2.5.subset.bed',
+	import.meta.url,
+);
+
+const defaultBedUrls: DefaultBedUrls = {
+	WGS: bedShuffled1Url.href,
+	WXS: bedIlluminaUrl.href,
+	ChipSeq: bedShuffled2Url.href,
+	'RNA-Seq': bedIlluminaUrl.href,
+};
+
+/** Lookup Default Bed File
+ * @param elasticDocument File Centric ElasticSearch document
+ * @returns bedFileUrl - string
+ */
+export const getBedFileUrl = (elasticDocument: FileDocument) => {
+	const fileStrategy = elasticDocument.analysis?.experiment?.experimentalStrategy;
+	const isValidStrategy = fileStrategy && fileStrategies.includes(fileStrategy);
+	const bedFileUrl = isValidStrategy ? defaultBedUrls[fileStrategy] : '';
+
+	return bedFileUrl;
 };
